Add ElasticsearchException to model Elasticsearch exceptions
This commit adds an ElasticsearchException struct that models
exceptions returned by Elasticsearch in the event of an error.

An exception() function is exposed on Response to deserialize the
response body into ElasticsearchException when the status code is
in the 400-599 range.

Consideration was made to exposing this on
the Error struct, performing the deserialization inside
`error_for_status_code`, but this introduces some awkwardness into
the API:

- `error_for_status_code` would ideally be async because the .json()
  call is async
- .json() consumes the response, which would not be possible for
  `error_for_status_code_ref` which has a shared reference to Response.

Additional details about an exception are collected in a map and exposed
through additional_details function.

Closes #112